Move FileUri to Domain
Location will move to Domain in the same step, and needs FileUri unchanged;
Domain has no outbound edges, so FileUri moves with it. Parser used FileUri
only inside parseFile to fill a TextDocument URI the parse flow never reads,
so the call drops rather than pulling a Parser -> Domain edge.
